Refactor decryption into separate class.
